+++ src/reprinter/index.ts
@@ -19,12 +19,28 @@ import { sortUnionTypeAnnotation } from "../sortUnionTypeAnnotation";
 // Utils
 import { isArray } from "util";
 import { endsWith } from "../common/string-utils";
+import { Logger, LoggerVerboseOption } from "../logger";
 
 export interface ReprinterOptions {
+  // Default "false". If true, prints out very verbose lines that sortier doesn't know how to handle so you can open Github issues about them
   isHelpMode?: boolean;
+
+  // Default "false". If true, sortier will run but not rewrite any files. Great for testing to make sure your code base doesn't have any weird issues before rewriting code.
   isTestRun?: boolean;
+
+  // Default "normal". This overrides isHelpMode if set.
+  //  - "quiet" - No console logs
+  //  - "normal" - General information (e.g. if sortier was unable to parse a file)
+  //  - "diagnostic" - All the above along with type information that sortier was unable to handle (great for opening bugs!)
+  logLevel?: "diagnostic" | "normal" | "quiet";
+
+  // Default undefined. The parser to use. If undefined, sortier will determine the parser to use based on the file extension
   parser?: "flow" | "typescript";
+
+  // Default "source". The order you wish to sort import statements. Source is the path the import comes from. First specifier is the first item imported.
   sortImportDeclarations?: "first-specifier" | "source";
+
+  // Default ["undefined", "null", "*", "object", "function"]. The order to sort object types when encountered.
   sortTypeAnnotations?: ("null" | "undefined" | "*" | "function" | "object")[];
 }
 
@@ -722,19 +738,20 @@ export class Reprinter {
   private printHelpModeInfo(item, fileContents: string) {
     if (this._options.isHelpMode === true) {
       if (!this._helpModeHasPrintedFilename) {
-        console.log("");
-        console.log(this._filename);
+        Logger.log(LoggerVerboseOption.Diagnostic, "");
+        Logger.log(LoggerVerboseOption.Diagnostic, this._filename);
       }
 
-      console.log(
-        " - " +
-          item.type +
-          " - " +
-          JSON.stringify(item.loc.start) +
-          " - " +
-          JSON.stringify(item.loc.end)
+      Logger.log(
+        LoggerVerboseOption.Diagnostic,
+        ` - ${item.type} - ${JSON.stringify(item.loc.start)} - ${JSON.stringify(
+          item.loc.end
+        )}`
       );
-      console.log(fileContents.substring(item.range[0], item.range[1]));
+      Logger.log(
+        LoggerVerboseOption.Diagnostic,
+        fileContents.substring(item.range[0], item.range[1])
+      );
     }
   }
 }